+++ tests/test_client.py
@@ -716,10 +716,10 @@ def test_collections_are_clients() -> None:
     catalog = Client.open(
         "https://planetarycomputer.microsoft.com/api/stac/v1/",
     )
-    item = catalog.get_collection("cil-gdpcir-cc-by").get_item(
-        "cil-gdpcir-NUIST-NESM3-ssp585-r1i1p1f1-day"
-    )
-    assert item
+    collection = catalog.get_collection("cil-gdpcir-cc-by")
+    assert collection is not None
+    item = collection.get_item("cil-gdpcir-NUIST-NESM3-ssp585-r1i1p1f1-day")
+    assert item is not None
     item.get_collection()
 
 
@@ -747,6 +747,8 @@ def test_fallback_strategy() -> None:
         "https://planetarycomputer.microsoft.com/api/stac/v1/",
     )
     col = client.get_collection("landsat-c2-l2")
+    assert col is not None
+
     item = next(col.get_items())
 
     item_href = item.self_href
@@ -786,19 +788,21 @@ def test_get_collection_fallback_strategy_for_static_catalogs() -> None:
 
     with pytest.warns(FallbackToPystac):
         area_1_1 = client.get_collection("area-1-1")
+        assert area_1_1 is not None
 
     with pytest.warns(FallbackToPystac):
         area_1_2 = client.get_collection("area-1-2")
+        assert area_1_2 is not None
 
     assert area_1_1.id == "area-1-1"
     assert area_1_2.id == "area-1-2"
 
 
-def test_get_collection_for_static_catalogs_raise_if_not_found() -> None:
+def test_get_collection_for_static_catalogs_returns_none_if_not_found() -> None:
     client = Client.from_file(str(TEST_DATA / "test-case-1" / "catalog.json"))
-    with pytest.raises(KeyError, match="not found on catalog"):
-        with pytest.warns(FallbackToPystac):
-            client.get_collection("country-1")
+    with pytest.warns(FallbackToPystac):
+        collection = client.get_collection("country-1")
+        assert collection is None
 
 
 def test_get_collection_raises_if_collection_id_is_empty() -> None:
@@ -808,6 +812,13 @@ def test_get_collection_raises_if_collection_id_is_empty() -> None:
 
     with pytest.raises(ValueError, match="A `collection_id` must be provided"):
         client.get_collection(None)
+
+
+@pytest.mark.vcr
+def test_get_collection_returns_none_if_not_found() -> None:
+    client = Client.open(STAC_URLS["EARTH-SEARCH"])
+    collection = client.get_collection("foo")
+    assert collection is None
 
 
 @pytest.mark.vcr